Move GetMimeType extension to primitives
