+++ src/CommunityToolkit.Maui.MediaElement/Views/MediaManager.windows.cs
@@ -3,6 +3,7 @@ using System.Numerics;
 using CommunityToolkit.Maui.Core.Primitives;
 using CommunityToolkit.Maui.Extensions;
 using CommunityToolkit.Maui.Views;
+using CommunityToolkit.Maui.Core.Primitives;
 using Microsoft.Extensions.Logging;
 using Microsoft.Maui;
 using Microsoft.Maui.Controls;
